### 2022-10-27 #108 @whedon
[PRE REVIEW]: MedPipe3D.jl: GPU accelerated medical image segmentation framework
**Submitting author:** @jakubMitura14(<a href="http://orcid.org/0000-0003-1823-6823">Jakub Mitura</a>)
**Repository:** <a href=" https://github.com/jakubMitura14/MedPipe3D.jl" target ="_blank"> https://github.com/jakubMitura14/MedPipe3D.jl</a>
**Version:**

Thanks for submitting your paper to JCON @https://github.com/jakubMitura14. **Currently, there isn't an JCON editor assigned** to your paper.
