### 2023-03-12 #108 @jakubMitura14
[JuliaCon2022Article.pdf](https://github.com/JuliaCon/proceedings-review/files/10950040/JuliaCon2022Article.pdf)

pdf is here and link to overleaf below - I do not know why whedon is unable to compile
https://www.overleaf.com/3256953646dpzjctndcdgb
