### 2023-03-12 #108 @mkitti
Whedon is probably looking for a paper.md file.
https://joss.readthedocs.io/en/latest/editorial_bot.html?highlight=paper.md#compiling-papers
