### 2023-04-20 #127 @whedon
[REVIEW]: MedPipe3D.jl: GPU accelerated medical image segmentation framework
**Submitting author:** @jakubMitura14 (<a href="http://orcid.org/0000-0003-1823-6823">Jakub Mitura</a>)
**Repository:**
**Branch with paper.md** (empty if default branch):
**Version:**

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the <a href=" https://github.com/jakubMitura14/MedPipe3D.jl" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

- [ ] **Repository:** Is the source code for this software available at the <a target="_blank" href=" https://github.com/jakubMitura14/MedPipe3D.jl">repository url</a>?

- [ ] **Repository:** Is the source code for this software available at the <a target="_blank" href=" https://github.com/jakubMitura14/MedPipe3D.jl">repository url</a>?
