### 2023-04-28 #127 @jakubMitura14
already compiled pdf is in paper folder
https://github.com/jakubMitura14/MedPipe3D.jl/blob/master/paper/test.pdf
