### 2023-09-07 #127 @matbesancon
@jakubMitura14 did you modify something on the `paper` folder compared to the submission template?
