+++ Project.toml
@@ -10,8 +10,8 @@ MedEval3D = "07fe485e-7252-451c-a484-d20a4cdbbf79"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 
 [compat]
-ColorTypes = "0.11.0,0.10.12"
-HDF5 = "0.16.5"
-MedEval3D = "0.1.0"
+ColorTypes = "0.11.4"
+HDF5 = "0.17.1"
+MedEval3D = "0.2.0"
 PythonCall = "0.8.0"
-julia = "1.6"
+julia = "1.9"
