### 2024-04-10 #127 @luraess
> Yes

I am thus finalising the withdrawal of the paper.
